move platform code to separate modules
