+++ src/lib.rs
@@ -1,33 +1,12 @@
+#[cfg(target_os = "macos")]
+mod macos;
+
+#[cfg(target_os = "windows")]
+mod windows;
+
 use std::fmt;
 
-//============================================================================//
-// External platform specific API functions.                                  //
-//============================================================================//
-// MacOS.
-#[cfg(target_os = "macos")]
-#[link(name = "login", kind = "framework")]
-extern "C" {
-    fn SACLockScreenImmediate();
-}
-
-// Windows.
 pub type Win32ErrorCode = u32;
-
-#[cfg(target_os = "windows")]
-#[link(name = "user32")]
-extern "system" {
-    fn LockWorkStation() -> i32;
-}
-
-#[cfg(target_os = "windows")]
-#[link(name = "Kernel32")]
-extern "system" {
-    fn GetLastError() -> Win32ErrorCode;
-}
-
-//============================================================================//
-// Lock screen crate implementation.                                          //
-//============================================================================//
 type Result<T> = std::result::Result<T, Error>;
 
 /// Error information explaining why the screen couldn't be locked.
@@ -53,23 +32,14 @@ impl fmt::Display for Error {
 /// Locks the computer screen by hiding the current desktop, and requiring
 /// the user to re-enter their password before continuing.
 pub fn lock_screen() -> Result<()> {
-    // TODO(scott): Linux lock screen - write code to invoke a list of programs.
-
     #[cfg(target_os = "macos")]
-    unsafe {
-        SACLockScreenImmediate();
-        Ok(())
-    }
+    return crate::macos::lock_screen_mac();
 
     #[cfg(target_os = "windows")]
-    unsafe {
-        if 0 == LockWorkStation() {
-            Err(Error::Win32(GetLastError()))
-        } else {
-            Ok(())
-        }
-    }
+    return crate::windows::lock_screen_windows();
+
+    // TODO(scott): Add linux support.
 
     #[cfg(not(any(target_os = "windows", target_os = "macos")))]
-    Err(Error::UnsupportedPlatform)
+    return Err(Error::UnsupportedPlatform);
 }
